move image proxy tests to the right file
