+++ tests/Filter/FilterTest.php
@@ -90,134 +90,4 @@ class FilterTest extends PHPUnit_Framework_TestCase
         $f->setConfig($config);
         $this->assertEquals('<p>Testboo</p>', $f->execute());
     }
-
-    public function testNoImageProxySet()
-    {
-        $f = Filter::html('<p>Image <img src="/image.png" alt="My Image"/></p>', 'http://foo');
-
-        $this->assertEquals(
-            '<p>Image <img src="http://foo/image.png" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyWithHTTPLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-
-        $f = Filter::html('<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://myproxy/?url='.rawurlencode('http://localhost/image.png').'" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyWithHTTPSLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-
-        $f = Filter::html('<p>Image <img src="https://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://myproxy/?url='.rawurlencode('https://localhost/image.png').'" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyLimitedToUnknownProtocol()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-        $config->setFilterImageProxyProtocol('tripleX');
-
-        $f = Filter::html('<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyLimitedToHTTPwithHTTPLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-        $config->setFilterImageProxyProtocol('http');
-
-        $f = Filter::html('<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://myproxy/?url='.rawurlencode('http://localhost/image.png').'" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyLimitedToHTTPwithHTTPSLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-        $config->setFilterImageProxyProtocol('http');
-
-        $f = Filter::html('<p>Image <img src="https://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="https://localhost/image.png" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyLimitedToHTTPSwithHTTPLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-        $config->setFilterImageProxyProtocol('https');
-
-        $f = Filter::html('<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://localhost/image.png" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testImageProxyLimitedToHTTPSwithHTTPSLink()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyUrl('http://myproxy/?url=%s');
-        $config->setFilterImageProxyProtocol('https');
-
-        $f = Filter::html('<p>Image <img src="https://localhost/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="http://myproxy/?url='.rawurlencode('https://localhost/image.png').'" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
-
-    public function testsetFilterImageProxyCallback()
-    {
-        $config = new Config;
-        $config->setFilterImageProxyCallback(function ($image_url) {
-            $key = hash_hmac('sha1', $image_url, 'secret');
-            return 'https://mypublicproxy/'.$key.'/'.rawurlencode($image_url);
-        });
-
-        $f = Filter::html('<p>Image <img src="/image.png" alt="My Image"/></p>', 'http://foo');
-        $f->setConfig($config);
-
-        $this->assertEquals(
-            '<p>Image <img src="https://mypublicproxy/4924964043f3119b3cf2b07b1922d491bcc20092/'.rawurlencode('http://foo/image.png').'" alt="My Image"/></p>',
-            $f->execute()
-        );
-    }
 }
